fix: 40 bug fixes across nn, data, optimizers, schedulers, io, and autograd
- layers.py: Fix Flatten off-by-one (removed end+1), add train()/eval() propagation
  to Python child layers, raise KeyError on missing/unexpected load_state_dict keys
- data.py: Remove invalid np.arange out= param, validate num_samples bounds
- tensor.py: Add dtype error (not yet supported), validate shape vs flatten size,
  handle tuples in _flatten
- schedulers.py: Make ReduceLROnPlateau.metric optional, add state_dict/load_state_dict
  and get_last_lr to base, fix _set_lr to update param_groups directly
- optimizers.py: Add norm_type > 0 validation to clip_grad_norm_
- parallel.py: Fix integer truncation in weighted splits, add zero-batch validation,
  use consistent weighted loss averaging, implement actual device placement
- callbacks.py: Move restored weights to model device before load_state_dict
- io/onnx.py: Fix transB default 1→0, apply Gemm alpha/beta coefficients
- models/base.py: Reconstruct model from metadata on load, save model class/config
- module.py: Fix to_gpu type annotation to device_id: int
- nn.py: Add GELU naming comment
- typing.py: Add gpu/gpu:N to Device type
- src/python/ops.rs: Replace checkpoint identity no-op with real gradient
  checkpointing, wrap bce_with_logits/huber_loss with autograd
- src/autograd/losses.rs: Fix CE targets dtype (f32/i64 mismatch), remove BCE
  infinite-loss→0 guard, add target to MSELossBackward.inputs, fix
  EmbeddingBackward.num_inputs=2
- src/kernels/cpu/losses.rs: Remove BCE is_finite guard
- src/autograd/reductions.rs: Fix LogSoftmaxBackward dim usize→i32 cast
- src/kernels/cpu/reductions.rs: Fix mean_kernel None→reduce-all-dims
- src/tensor/ops.rs: Fix op alias naming

+++ src/tensor/ops.rs
@@ -1173,7 +1173,7 @@ impl Tensor {
             let edges = autograd::make_edge(self);
             let backward = Arc::new(autograd::LogSoftmaxBackward::new(
                 output.clone(),
-                dim as usize,
+                dim as i32,
                 edges,
             ));
             Self::attach_grad_fn(output, backward)
